refactor: Streamline background task handler and registry structure by removing unused methods and consolidating task execution logic
