+++ src/ai/backend/common/bgtask/task/registry.py
@@ -1,4 +1,5 @@
-from typing import Any, Mapping
+from abc import ABC, abstractmethod
+from typing import Any, Generic, Mapping, override
 
 from ai.backend.common.exception import (
     BgtaskNotRegisteredError,
@@ -9,26 +10,58 @@ from .base import (
     BaseBackgroundTaskArgs,
     BaseBackgroundTaskHandler,
     BaseBackgroundTaskResult,
-    TaskExecutor,
+    TFunctionArgs,
 )
 
 
+class _TaskExecutor(ABC):
+    @abstractmethod
+    async def revive_task(self, args: Mapping[str, Any]) -> BaseBackgroundTaskResult:
+        """
+        Revive the background task with the provided arguments.
+        This method should be implemented by subclasses to provide
+        the specific revival logic.
+        """
+        raise NotImplementedError("Subclasses must implement this method")
+
+    @abstractmethod
+    async def execute_new_task(self, args: BaseBackgroundTaskArgs) -> BaseBackgroundTaskResult:
+        """
+        Execute the background task with the provided reporter and arguments.
+        This method should be implemented by subclasses to provide
+        the specific execution logic.
+        """
+        raise NotImplementedError("Subclasses must implement this method")
+
+
+class _TaskDefinition(_TaskExecutor, Generic[TFunctionArgs]):
+    _handler: BaseBackgroundTaskHandler[TFunctionArgs]
+
+    def __init__(self, handler: BaseBackgroundTaskHandler[TFunctionArgs]) -> None:
+        self._handler = handler
+
+    @override
+    async def revive_task(self, args: Mapping[str, Any]) -> BaseBackgroundTaskResult:
+        args_instance = self._handler.args_type().from_redis_json(args)
+        return await self._handler.execute(args_instance)
+
+    @override
+    async def execute_new_task(self, args: BaseBackgroundTaskArgs) -> BaseBackgroundTaskResult:
+        if not isinstance(args, self._handler.args_type()):
+            raise TypeError(f"Expected args of type {self._handler.args_type()}, got {type(args)}")
+        return await self._handler.execute(args)
+
+
 class BackgroundTaskHandlerRegistry:
-    _registry: dict[TaskName, BaseBackgroundTaskHandler]
-    _executor_registry: dict[TaskName, TaskExecutor]
+    _executor_registry: dict[TaskName, _TaskExecutor]
 
     def __init__(self) -> None:
-        self._registry = {}
         self._executor_registry = {}
 
-    def register(self, task: BaseBackgroundTaskHandler) -> None:
-        self._registry[task.name()] = task
-
-    def get_task(self, name: TaskName) -> BaseBackgroundTaskHandler:
-        try:
-            return self._registry[name]
-        except KeyError:
-            raise BgtaskNotRegisteredError
+    def register(self, handler: BaseBackgroundTaskHandler) -> None:
+        self._executor_registry[handler.name()] = _TaskDefinition(
+            handler=handler,
+        )
 
     async def revive_task(
         self, name: TaskName, args: Mapping[str, Any]
